+++ Tests/Functional/Command/BuildQueueCommandTest.php
@@ -62,7 +62,7 @@ class BuildQueueCommandTest extends AbstractCommandTests
         self::assertContains($expectedOutput, $commandOutput);
         self::assertEquals(
             $expectedCount,
-            $this->queueRepository->countAll()
+            $this->queueRepository->findAll()->count()
         );
     }